requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
curl-cffi==0.5.10
disnake==2.9.1
freeGPT==1.3.4
frozenlist==1.4.0
idna==3.6
llvmlite==0.42.0
memory-profiler==0.61.0
multidict==6.0.4
mutagen==1.47.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
psutil==5.9.8
pycparser==2.21
pycryptodomex==3.19.0
PyNaCl==1.5.0
requests==2.31.0
urllib3==2.1.0
websockets==12.0
yarl==1.9.4
yt-dlp==2023.12.6.232721.dev0
